Fix/libmagic fix (#104)
* ongoing tests

* ongoing fix

* added librt for linking

* fixing…

* build (docker): change base image

* build (travis): install libmagic with homebrew on macOS

* ongoing tests

* ongoing fix

* added librt for linking

* fixing…

* build (docker): change base image

* build (travis): install libmagic with homebrew on macOS

* Remove „register“ keyword

* Overflow with large files

* Refactor: image width/height as size_t

Some calculation (e.g. nx*ny*nc*bps/8) resulted in a overflow for large images. Thus changed image dimensions to size_t (aka unsigned long).
Attention: Delete the cache since the cache format is not compatible with the previous version!

* compile warnings fixes

* ongoinmg conversion to size_t

* ongoing…

* Bugfix with large XMP & JPEG

* Fix with photometric and rights of jpeg

* ongoing tests

* ongoing fix

* added librt for linking

* fixing…

* build (docker): change base image

* build (travis): install libmagic with homebrew on macOS

* added librt for linking

* fixing…

* Remove „register“ keyword

* Overflow with large files

* Refactor: image width/height as size_t

Some calculation (e.g. nx*ny*nc*bps/8) resulted in a overflow for large images. Thus changed image dimensions to size_t (aka unsigned long).
Attention: Delete the cache since the cache format is not compatible with the previous version!

* compile warnings fixes

* ongoinmg conversion to size_t

* ongoing…

* Bugfix with large XMP & JPEG

* Removed one last NULL

* fix (build): Download and compile librtmp.

* Add rtmp to CMakeLists.txt.

* Disable building rtmp for now.

* Remove rtmp, let’s see if we can do without it.

+++ include/SipiIO.h
@@ -62,7 +62,7 @@ namespace Sipi {
         * \param[out] width Width of the image in pixels
         * \param[out] height Height of the image in pixels
         */
-        virtual bool getDim(std::string filepath, int &width, int &height) = 0;
+        virtual bool getDim(std::string filepath, size_t &width, size_t &height) = 0;
 
         /*!
          * Write an image for a file using the given file format implemented by the subclass
